+++ src/main/java/com/cells/Tags.java
@@ -5,5 +5,5 @@ public class Tags {
 
     public static final String MODID = "cells";
     public static final String MODNAME = "C.E.L.L.S.";
-    public static final String VERSION = "0.2.1";
+    public static final String VERSION = "0.2.2";
 }